moved code to services and cleaned up
moved code to services and cleaned up
